Remove import cycle in `src/plugins/r2`
This probably wasn't a problem, but generally it's not great to have
cyclic dependencies. For future reference, this was found with
`npx madge --circular --extensions ts packages/miniflare/src`.
